Change ReservationDeleteModal into container
Makes adding the modal to different parts of the application
easier.
Refs #173.
