+++ app/containers/ReservationsList.js
@@ -6,25 +6,20 @@ import { bindActionCreators } from 'redux';
 import { pushState } from 'redux-router';
 
 import {
-  closeDeleteReservationModal,
   openDeleteReservationModal,
   selectReservationToDelete,
   selectReservationToEdit,
 } from 'actions/uiActions';
-import {
-  deleteReservation,
-  fetchReservations,
- } from 'actions/reservationActions';
+import { fetchReservations } from 'actions/reservationActions';
 import { fetchResources } from 'actions/resourceActions';
 import { fetchUnits } from 'actions/unitActions';
-import DeleteModal from 'components/reservation/DeleteModal';
+import ReservationDeleteModal from 'containers/ReservationDeleteModal';
 import ReservationsListItem from 'components/reservation/ReservationsListItem';
 import reservationsListSelector from 'selectors/containers/reservationsListSelector';
 
 export class UnconnectedReservationsList extends Component {
   constructor(props) {
     super(props);
-    this.handleDelete = this.handleDelete.bind(this);
     this.renderReservationsListItem = this.renderReservationsListItem.bind(this);
   }
 
@@ -32,14 +27,6 @@ export class UnconnectedReservationsList extends Component {
     this.props.actions.fetchResources();
     this.props.actions.fetchUnits();
     this.props.actions.fetchReservations();
-  }
-
-  handleDelete() {
-    const { actions, reservationsToDelete } = this.props;
-
-    _.forEach(reservationsToDelete, (reservation) => {
-      actions.deleteReservation(reservation);
-    });
   }
 
   renderReservationsListItem(reservation) {
@@ -67,13 +54,8 @@ export class UnconnectedReservationsList extends Component {
 
   render() {
     const {
-      actions,
-      deleteReservationModalIsOpen,
-      isDeletingReservations,
       isFetchingReservations,
       reservations,
-      reservationsToDelete,
-      resources,
     } = this.props;
 
     return (
@@ -83,14 +65,7 @@ export class UnconnectedReservationsList extends Component {
             <ul className="reservations-list">
               {_.map(reservations, this.renderReservationsListItem)}
             </ul>
-            <DeleteModal
-              onClose={actions.closeDeleteReservationModal}
-              onConfirm={this.handleDelete}
-              isDeleting={isDeletingReservations}
-              reservationsToDelete={reservationsToDelete}
-              resources={resources}
-              show={deleteReservationModalIsOpen}
-            />
+            <ReservationDeleteModal />
           </div>
         ) : (
           <p>Sinulla ei vielä ole yhtään varausta.</p>
@@ -102,19 +77,14 @@ export class UnconnectedReservationsList extends Component {
 
 UnconnectedReservationsList.propTypes = {
   actions: PropTypes.object.isRequired,
-  deleteReservationModalIsOpen: PropTypes.bool.isRequired,
-  isDeletingReservations: PropTypes.bool.isRequired,
   isFetchingReservations: PropTypes.bool.isRequired,
   reservations: PropTypes.array.isRequired,
-  reservationsToDelete: PropTypes.array.isRequired,
   resources: PropTypes.object.isRequired,
   units: PropTypes.object.isRequired,
 };
 
 function mapDispatchToProps(dispatch) {
   const actionCreators = {
-    closeDeleteReservationModal,
-    deleteReservation,
     fetchReservations,
     fetchResources,
     fetchUnits,
